test+fix(equipment): migration idempotency + shot round-trip tests; rpm type unify
Address the PR review's test-coverage gaps and the type-consistency nit:
- tst_dbmigration: v22_reinitDoesNotDuplicatePackages (re-init must not duplicate
  the default package — the migration data step is non-idempotent, guarded only by
  the version gate) and v22_shotEquipmentRpmRoundTrip (guards the positional
  equipment_id/rpm read at SELECT indices 39/40 + NULL->0 mapping).
- tst_equipment: split-heuristic edges (rpm-word-no-digits, non-trailing rpm,
  trailing whitespace) and mergeAndNameEdges (unused-source merge hard-deletes;
  editing into a RETIRED identity forks instead of resurrecting; name derived
  from a partial identity).
- The idempotency test surfaced a real bug: migration 22 read grinder settings
  via a bare QSettings() (depends on QCoreApplication org/app being set — the
  scope-mismatch class migration 16 fixed). Now reads the explicit
  ("DecentEspresso","DE1Qt") scope SettingsDye writes to.
- Unify the rpm type to qint64 across the shot structs (ShotRecord/ShotSaveData/
  ShotProjection/ShotMetadata) to match CoffeeBag/EquipmentPackage and stop the
  qint64->int narrowing at the bag<->shot boundary.

Full suite green (2390).

+++ src/history/shothistory_types.h
@@ -48,7 +48,7 @@ struct ShotRecord {
     QString grinderBurrs;
     QString grinderSetting;
     qint64 equipmentId = 0;  // FK -> equipment_packages.id (add-equipment-packages); 0 = none
-    int rpm = 0;             // grinder rpm dial-in; 0 = unset
+    qint64 rpm = 0;          // grinder rpm dial-in; 0 = unset
     double drinkTds = 0;
     double drinkEy = 0;
     QString espressoNotes;
@@ -236,7 +236,7 @@ struct ShotSaveData {
     QString grinderBurrs;
     QString grinderSetting;
     qint64 equipmentId = 0;  // FK -> equipment_packages.id (add-equipment-packages); 0 = none
-    int rpm = 0;             // grinder rpm dial-in; 0 = unset
+    qint64 rpm = 0;          // grinder rpm dial-in; 0 = unset
     double drinkTds = 0;
     double drinkEy = 0;
     int espressoEnjoyment = 0;
